Adding change.seq, for recording order. Reference:
Limits In SQLite, 13. Maximum Number Of Rows In A Table
https://sqlite.org/limits.html

The theoretical maximum number of rows in a table is 264 (18446744073709551616 or about 1.8e+19). This limit is unreachable since the maximum database size of 281 terabytes will be reached first. A 281 terabytes database can hold no more than approximately 2e+13 rows, and then only if there are no indices and if each row contains very little data.